refactor(codegen): split prop-object emission into its own module
Move `generate_props_object` / `generate_props_object_inner` out of
`generate.rs` into a new `props/object.rs` (#1756 groundwork).

`generate.rs` was at 589 lines — over the source-length guard — which left no
room to route prop emission through the Rendu plate in place. Splitting the
props-object emission into its own module drops `generate.rs` to 331 and gives
both files headroom for the staged emission reroutes.

Pure code movement — no logic change. Verified byte-equivalent across the full
Atelier snapshot suites: `vize_atelier_dom`, `vize_atelier_sfc` (528), and
`vize_atelier_ssr` (76) all pass with no snapshot changes.
